# colblk: add PrefixBytes.AppendAt,SetNextAt