**colblk: add PrefixBytes.AppendAt**

Add a new PrefixBytes.AppendAt method that appends the slice at the i'th index
to the provided destination slice. This commit also adjusts the
BenchmarkPrefixBytes microbenchmark to access slices sequentially.

```
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/pebble/sstable/colblk
                                     │ prefixbytes-old.txt │
                                     │       sec/op        │
PrefixBytes/alphaLen=2/iteration-10            22.61n ± 0%
PrefixBytes/alphaLen=5/iteration-10            23.18n ± 2%
PrefixBytes/alphaLen=26/iteration-10           23.62n ± 0%
geomean                                        23.13n
```

**colblk: use pointer receiver for RawBytes, PrefixBytes**
```
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/pebble/sstable/colblk
                                     │ prefixbytes-old.txt │         prefixbytes-new.txt         │
                                     │       sec/op        │   sec/op     vs base                │
PrefixBytes/alphaLen=2/iteration-10            22.61n ± 0%   18.99n ± 1%  -16.03% (p=0.000 n=20)
PrefixBytes/alphaLen=5/iteration-10            23.18n ± 2%   19.68n ± 4%  -15.12% (p=0.000 n=20)
PrefixBytes/alphaLen=26/iteration-10           23.62n ± 0%   19.68n ± 0%  -16.68% (p=0.000 n=20)
geomean                                        23.13n        19.44n       -15.95%
```

**colblk: inline SharedPrefix in PrefixBytes.AppendAt**

```
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/pebble/sstable/colblk
                                     │ prefixbytes-old.txt │        prefixbytes-new.txt         │
                                     │       sec/op        │   sec/op     vs base               │
PrefixBytes/alphaLen=2/iteration-10            18.99n ± 1%   17.73n ± 5%  -6.61% (p=0.000 n=20)
PrefixBytes/alphaLen=5/iteration-10            19.68n ± 4%   18.65n ± 1%  -5.18% (p=0.000 n=20)
PrefixBytes/alphaLen=26/iteration-10           19.68n ± 0%   18.80n ± 2%  -4.42% (p=0.000 n=20)
geomean                                        19.44n        18.39n       -5.41%
```

**colblk: add PrefixBytes.SetNextAt**

Add a PrefixBytes.SetNextAt that may be used when materializing the next slice
in a PrefixBytes into a buffer known to contain the previous slice. SetNextAt
can take advantage of the previous slice to always avoid copying the shared
prefix, often avoid copying the bundle prefix and occasionally avoid copying
the row suffix.

```
goos: darwin
goarch: arm64
pkg: github.com/cockroachdb/pebble/sstable/colblk
                                     │ prefixbytes-old.txt │         prefixbytes-new.txt         │
                                     │       sec/op        │   sec/op     vs base                │
PrefixBytes/alphaLen=2/iteration-10           17.735n ± 5%   7.811n ± 1%  -55.96% (p=0.000 n=20)
PrefixBytes/alphaLen=5/iteration-10           18.655n ± 1%   7.717n ± 1%  -58.63% (p=0.000 n=20)
PrefixBytes/alphaLen=26/iteration-10          18.805n ± 2%   7.600n ± 1%  -59.58% (p=0.000 n=20)
geomean                                        18.39n        7.709n       -58.08%
```